chore(일부수정): 사소한 수정

+++ src/main/java/project/freemates2/api/place/domain/Place.java
@@ -41,9 +41,9 @@ public class Place extends BaseEntity {
   @Builder.Default
   private Long reviewCount = 0L;
 
-  private int lat; // 위도
+  private Double lat; // 위도
 
-  private int lng; // 경도
+  private Double lng; // 경도
 
   private String phone;
